extract `Conditions` (for the WHERE clause) and use where applicable; remove `ProjectionQuery` and refactor per #45; documentation updates and upgrade notes. (more refactoring and MySQL-specific tests needed for feature parity with MySQL-specific LIMIT/JOIN clauses in UPDATE/DELETE queries.)
